+++ package/telegram_client/lib/telegram_bot_api_server/ffi/telegram_bot_api_server.dart
@@ -1,0 +1,2 @@
+export "telegram_bot_api_server_io.dart"
+    if (dart.library.html) "telegram_bot_api_server_web.dart";